登录界面和Money输入的BUG修复

+++ app/src/main/java/com/example/wangchang/wda1150_xwk1151/AddAccount_forFloatingActivity.java
@@ -42,7 +42,7 @@ public class AddAccount_forFloatingActivity extends AppCompatActivity {
     String name;
     String type;
     String month;
-    double money;
+    Float money;
 
     AccountBeen new_account;
